Implemented ASSIGNMENT
- Implemented assignment end-to-end
- Corrected a bug with FLOAT analysis (const flt := -.5E-)

+++ pika/src/parseTree/nodeTypes/FloatingConstantNode.java
@@ -17,7 +17,7 @@ public class FloatingConstantNode extends ParseNode {
 ////////////////////////////////////////////////////////////
 // attributes
 	
-	public float getValue() {
+	public double getValue() {
 		return floatingToken().getValue();
 	}
